Moved src/stan/services/mcmc/print_progress.hpp to src/stan/services/sample/progress.hpp
