+++ src/test/unit/services/sample/progress_test.cpp
@@ -1,8 +1,8 @@
-#include <stan/services/mcmc/print_progress.hpp>
+#include <stan/services/sample/progress.hpp>
 #include <gtest/gtest.h>
 #include <sstream>
 
-TEST(StanUi, print_progress) {
+TEST(StanUi, progress) {
   std::stringstream out;
   int m;
   int start;
@@ -21,16 +21,16 @@ TEST(StanUi, print_progress) {
 
   out.str("");
   m = 0;
-  stan::services::mcmc::print_progress(m, start, finish, refresh, warmup,
-                               prefix, suffix, out);
+  stan::services::sample::progress(m, start, finish, refresh, warmup,
+                                   prefix, suffix, out);
   EXPECT_EQ("Iteration:  1 / 100 [  1%]  (Warmup)", out.str());
 
   out.str("");
   m = 1;
   prefix = "\r";
   suffix = "\n";
-  stan::services::mcmc::print_progress(m, start, finish, refresh, warmup,
-                               prefix, suffix, out);
+  stan::services::sample::progress(m, start, finish, refresh, warmup,
+                                   prefix, suffix, out);
   EXPECT_EQ("\rIteration:  2 / 100 [  2%]  (Warmup)\n", out.str());
 
   out.str("");
@@ -38,8 +38,8 @@ TEST(StanUi, print_progress) {
   warmup = false;
   prefix = "";
   suffix = "";
-  stan::services::mcmc::print_progress(m, start, finish, refresh, warmup,
-                               prefix, suffix, out);
+  stan::services::sample::progress(m, start, finish, refresh, warmup,
+                                   prefix, suffix, out);
   EXPECT_EQ("Iteration:  3 / 100 [  3%]  (Sampling)", out.str());
 }
